Update AutofacExtensions.cs
Fixed a bug where calling As<T> on the result of RegisterAndActivate resulted in overriding the original Register call, and when StartableBootstrap tried to resolve the actual type, the container threw an exception. By calling As<T> in RegisterAndActivate, subsequent calls to As<..> will add to the registered underlying type rather than hide it.